refactor: reorganize question and category entities; update imports and remove unused category API files; update .env file
