+++ .env
@@ -1,10 +1,1 @@
 VITE_API_KEY='jWbpWSe3AJbIkDyneQqBO6xNP5rOlqd2hO8hhwug'
-VITE_API_BASE_URL='https://quizapi.io/api/v1/'
-
-VITE_API_ENDPOINT_GET_CATEGORIES='categories'
-VITE_API_ENDPOINT_GET_TAGS='tags'
-VITE_API_ENDPOINT_QUESTIONS='questions'
-
-VITE_API_DIFFICULTY_EASY='easy'
-VITE_API_DIFFICULTY_MEDIUM='medium'
-VITE_API_DIFFICULTY_HARD='hard'
