+++ src/entities/Question/api/getQuestions/index.ts
@@ -1,0 +1,8 @@
+export { getQuestions } from './getQuestions';
+export type {
+  QuestionModel,
+  RequestQuestionsParams,
+  Difficulty,
+  Answers,
+  IsCorrect,
+} from './getQuestions.types';
