+++ src/entities/Question/index.ts
@@ -1,0 +1,10 @@
+export {
+  getQuestions,
+  type QuestionModel,
+  type RequestQuestionsParams,
+  type Difficulty,
+  type Answers,
+  type IsCorrect,
+} from './api/getQuestions';
+
+export { limitQuestions } from './model/config';
