+++ src/entities/Question/lib/endpoints.ts
@@ -1,0 +1,1 @@
+export const ENDPOINT_QUESTIONS = '/questions';
