+++ src/entities/Question/model/config.ts
@@ -1,0 +1,1 @@
+export const limitQuestions = 10;
