--- src/entities/category/api/getCategories.types.ts
@@ -1,8 +1,0 @@
-export interface CategoryModel {
-  id: number;
-  name: string;
-}
-
-export interface CategoriesDataResponse {
-  data: CategoryModel;
-}
