--- src/entities/category/lib/constants/index.ts
@@ -1,1 +1,0 @@
-export { ENDPOINT_CATEGORIES } from './endpoints';
