--- src/entities/questions/index.ts
@@ -1,9 +1,0 @@
-export { getQuestions } from './questions.api';
-export type {
-  QuestionModel,
-  RequestQuestionsParams,
-  Difficulty,
-  Answers,
-  IsCorrect,
-} from './questions.types';
-export { limitQuestions } from './questions.const';
