--- src/entities/questions/questions.const.ts
@@ -1,2 +1,0 @@
-export const ENDPOINT_QUESTIONS = import.meta.env.VITE_API_ENDPOINT_QUESTIONS;
-export const limitQuestions = 10;
