+++ src/entities/Сategory/api/getCategories/getCategories.types.ts
@@ -1,0 +1,4 @@
+export interface CategoryModel {
+  id: number;
+  name: string;
+}
