+++ src/entities/Сategory/api/getCategories/index.ts
