+++ src/entities/Сategory/index.ts
