+++ src/entities/Сategory/lib/endpoints.ts
@@ -1,0 +1,1 @@
+export const ENDPOINT_CATEGORIES = '/categories';
