+++ src/pages/game/game-page.ui.tsx
@@ -1,11 +1,16 @@
-import { Box, Button, CircularProgress, Typography } from '@mui/material';
-import { useEffect } from 'react';
 import { Link, useParams } from 'react-router-dom';
-import { Difficulty, limitQuestions } from '@/entities/questions';
+import { useEffect } from 'react';
+import { Box, Button, CircularProgress, Typography } from '@mui/material';
+
+import { useAppStore } from '@/app/store';
+
+import { Difficulty, limitQuestions } from '@/entities/Question';
+
 import { AnswerList } from '@/widgets/answer';
+
 import { NextQuestionButton } from '@/features/answer';
+
 import { pathKeys } from '@/shared/lib/react-router';
-import { useAppStore } from '@/app/store/store';
 
 export function GamePage() {
   const numberOfAnswers = useAppStore((state) => state.answers.numberOfAnswers);
